+++ Cargo.toml
@@ -1,36 +1,19 @@
 [package]
 name = "arzte"
-version = "0.0.3-alpha"
+version = "0.1.0"
 authors = ["Faey Hoffmeyer <hoffmeyer25@gmail.com>"]
 publish = false
 build = "build.rs"
 edition = "2018"
 
 [dependencies]
-serenity = { git = "https://github.com/Arzte/serenity.git", branch = "current"}
-config = "0.9.1"
-env_logger = "0.5.13"
-kankyo = "0.2.0"
-log = "0.4.6"
-rand = "0.6.3"
-chrono = "0.4.6"
-typemap = "0.3.3"
-sys-info = "0.5.6"
-built = "0.3.0"
-serde = "1.0.84"
-serde_derive = "1.0.84"
-serde_json = "1.0.34"
-reqwest = "0.9.5"
-meval = "0.2.0"
-sentry = "0.12.0"
-tempdir = "0.3.7"
-tar = "0.4.20"
-flate2 = "1.0.6"
-failure = "0.1.3"
-asciimath = { git =  "https://github.com/Arzte/asciimath-rs.git", branch = "patch-1" }
+env_logger = "0.6"
+kankyo = "0.3.0"
+log = "0.4"
+
+[dependencies.serenity]
+version = "0.6"
+features = ["cache", "voice", "http", "framework", "standard_framework", "rustls_backend"]
 
 [build-dependencies]
-built = "0.3.0"
-
-[replace]
-"openssl:0.9.24" = { git = "https://github.com/ishitatsuyuki/rust-openssl", branch = "0.9.x" }
+built = "0.3.1"
