+++ ci/before_deploy.sh
@@ -17,10 +17,8 @@ main() {
 
     test -f Cargo.lock || cargo generate-lockfile
 
-    # TODO Update this to build the artifacts that matter to you
     cargo build --bin arzte --release
 
-    # TODO Update this to package the right artifacts
     cp target/$TARGET/release/arzte $stage/
 
     cd $stage
