--- settings.example.toml
@@ -1,2 +1,0 @@
-token = "MTc1MDQ4MTc0MzMzOTg0NzY4.DxQwWQ.7TYu1RMBpUkTFlkxhYHsqR_89Nw"
-debug = "false"
