+++ temp/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "arzte"
+version = "0.0.3-alpha"
+authors = ["Faey Hoffmeyer <hoffmeyer25@gmail.com>"]
+publish = false
+build = "build.rs"
+edition = "2018"
+
+[dependencies]
+serenity = "0.6.2"
+config = "0.9.3"
+env_logger = "0.6.2"
+kankyo = "0.3.0"
+log = "0.4.7"
+rand = "0.7.0"
+chrono = "0.4.7"
+typemap = "0.3.3"
+sys-info = "0.5.7"
+built = "0.3.1"
+serde = "1.0.94"
+serde_derive = "1.0.94"
+serde_json = "1.0.40"
+reqwest = "0.9.18"
+meval = "0.2.0"
+sentry = "0.15.5"
+tempdir = "0.3.7"
+tar = "0.4.26"
+flate2 = "1.0.9"
+failure = "0.1.5"
+asciimath = "0.8.8"
+
+[build-dependencies]
+built = "0.3.1"
