+++ temp/src/commands/info.rs
