+++ temp/src/core/structs.rs
