+++ temp/src/core/utils.rs
